sample/mslearn-django-deployment-Azure/starter/requirements.txt
Django
pylint
autopep8
django-crispy-forms
whitenoise
psycopg2-binary
python-dotenv